+++ src/errors/index.js
@@ -3,4 +3,6 @@ module.exports = {
 	AntiVirusInitializationError: require("./AntiVirusInitializationError"),
 	AntiVirusPingError: require("./AntiVirusPingError"),
 	AntiVirusVersionError: require("./AntiVirusVersionError"),
+	AntiVirusMimeError: require("./AntiVirusMimeError"),
+	AntiVirusSizeError: require("./AntiVirusSizeError"),
 };